+++ meta/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to the Orbital Trading project will be documented in this file.
 
+## [28.00] - 2025-10-06
+- zoned travel cost updates
+- location supply/demand update
+- launch modal update
+- travel animation visual update
+- location quirks
+- bug fixes
+
 ## [27.50] - 2025-10-04
 - navigation bar visual update
 - bug fixes